fix: ensure logging imported in indent views
