+++ inventory/views/__init__.py
@@ -1,0 +1,27 @@
+from .api import (
+    ItemViewSet,
+    SupplierViewSet,
+    StockTransactionViewSet,
+    IndentViewSet,
+    IndentItemViewSet,
+    RecipeViewSet,
+    RecipeComponentViewSet,
+    PurchaseOrderViewSet,
+    PurchaseOrderItemViewSet,
+    GoodsReceivedNoteViewSet,
+    GRNItemViewSet,
+)
+
+__all__ = [
+    'ItemViewSet',
+    'SupplierViewSet',
+    'StockTransactionViewSet',
+    'IndentViewSet',
+    'IndentItemViewSet',
+    'RecipeViewSet',
+    'RecipeComponentViewSet',
+    'PurchaseOrderViewSet',
+    'PurchaseOrderItemViewSet',
+    'GoodsReceivedNoteViewSet',
+    'GRNItemViewSet',
+]
