+++ inventory/views/recipes.py
@@ -1,0 +1,42 @@
+from django.shortcuts import get_object_or_404, render
+from django.views import View
+from django.views.generic import TemplateView
+
+from ..models import Recipe, RecipeComponent
+
+
+class RecipesListView(TemplateView):
+    template_name = "inventory/recipes/list.html"
+
+    def get_context_data(self, **kwargs):
+        ctx = super().get_context_data(**kwargs)
+        ctx["recipes"] = Recipe.objects.all().order_by("name")
+        return ctx
+
+
+class RecipeDetailView(View):
+    template_name = "inventory/recipes/detail.html"
+
+    def get(self, request, pk: int):
+        recipe = get_object_or_404(Recipe, pk=pk)
+        components = RecipeComponent.objects.filter(parent_recipe=recipe).order_by(
+            "sort_order", "id"
+        )
+        return render(
+            request,
+            self.template_name,
+            {"recipe": recipe, "components": components},
+        )
+
+
+class RecipeComponentRowView(View):
+    template_name = "inventory/recipes/_component_row.html"
+
+    def get(self, request, pk: int):
+        recipe = get_object_or_404(Recipe, pk=pk)
+        index = int(request.GET.get("index", 0))
+        return render(
+            request,
+            self.template_name,
+            {"recipe": recipe, "index": index, "comp": None},
+        )
